+++ TDEngine2/include/TDEngine2.h
@@ -266,6 +266,7 @@
 #include "scene/components/CBaseLight.h"
 #include "scene/components/CDirectionalLight.h"
 #include "scene/components/CPointLight.h"
+#include "scene/components/CSpotLight.h"
 #include "scene/components/AudioComponents.h"
 #include "scene/components/CLODStrategyComponent.h"
 #include "scene/components/CObjIdComponent.h"