+++ TDEngine2/include/scene/IScene.h
@@ -107,6 +107,18 @@ namespace TDEngine2
 			TDE2_API virtual CEntity* CreatePointLight(const TColor32F& tint, F32 intensity, F32 range) = 0;
 
 			/*!
+				\brief The method is a helper that creates a new entity with PointLightComponent's attached to it
+
+				\param[in] tint A color of the light source
+				\param[in] intensity A power of contribution of the light source into the scene
+				\param[in] angle An angle of spotlight's cone in degrees
+
+				\return A pointer to a new created entity with CPointLightComponent component attached to it
+			*/
+
+			TDE2_API virtual CEntity* CreateSpotLight(const TColor32F& tint, F32 intensity, F32 angle) = 0;
+
+			/*!
 				\brief The method creates an entity which represents a skybox
 
 				\param[in, out] pResourceManager A pointer to IResourceManager implementation