+++ CHANGELOG.md
@@ -25,6 +25,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - GL_INVALID_OPERATION caused with incorrect glObjectLabel's calls was fixed.
 
+- Fix initialization of 3D texture under GLGraphicsContext.
+
 ## [0.6.25] 2024-03-03
 
 ### Added